parse: disambiguate assignments and declarations; declare FunDefs
AssignStmt's written as
   a, b = 3, 4
could be interpreted as an assignment or a declaration, without
the proper context. Using the rules of lexical scoping, this
context can be gotten by
  - checking that the lhs is an ident list.
  - checking the current scope for each identifier.
The language does not permit any mixing between old and new
variables.

Also declare function definitions in their containing scope if
they have a name.