types: implement type inference (#10)
This change exports the Infer function from the types package.
This function performs preliminary type inference and validation
for all expressions and identifiers in the syntax tree.

Type inference is based on the rules defined in types/HM.md for
Hindley-Milner type inference. However, implementation is split
into pre- and post-order traversals, in which partial types are
constructed on the way down, and fully-typed expressions are
reincorporated into a type on the way up. As such, only a single
full traversal is required. Finally, expressions and their
respective types are placed in the Types map listed in the Config
struct.

Updates #9.